Support running interpreter scripts
If the program starts with `#!` (a shebang or hash-bang), the kernel
will (almost always; depends if `BINFMT_SCRIPT` is enabled) determine
which interpreter to exec and pass the script along as the first
argument. In this case, the argument will be `/proc/self/fd/{}`, which
gets closed if MFD_CLOEXEC is set.

We now check for `#!` and only set MFD_CLOEXEC if it's not there.